Change Cart and Order
On Cart page we can create an Order, with special form. If User
authenticated, his phone and name will be show in the form. User
can change phone, order will be connected with this User anyway
(by user id). If User has no authentication and he write his phone,
order will be connected with this User if this phone exist in DB,
or will be create new User (last has bug).